cherrypick plugin restructuring from #301
