+++ src/plugins/filter.js
@@ -1,0 +1,8 @@
+import Vue from "vue";
+
+Vue.filter("striphtml", function(value) {
+	var div = document.createElement("div");
+	div.innerHTML = value;
+	var text = div.textContent || div.innerText || "";
+	return text;
+});
